Allow for multiple Franklin WH gates.
Also allows for setting a prefix to make it easier to determine
which gate is which when accessing sensors and switches.

As a byproduct of this, setting the (optional) unique_id allows
the frontend to manipulation the entity (such as adding friendly names).

Optional keys:

// Indicates whethere to add unique_ids based on the provided serial number.
// Default is False to preserve backwards compatibility.
use_sn: True/False

// Used for adding a prefix to entity names.
// Default if not supplied is to use the "FranklinWH" prefix for
// backwards compatibility.
prefix: "FranklinWH"